only user who login can view a post

+++ content-service/main.py
@@ -1,8 +1,9 @@
-from fastapi import FastAPI, Depends
+from fastapi import FastAPI, Depends,HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 import os
 from pydantic import BaseModel
 from content_queries import ContentQueries
+from auth import authenticator
 
 
 app = FastAPI()
@@ -36,12 +37,13 @@ class PostIn(BaseModel):
     title: str
     description: str
     subraddit: str
-    user: str
+    user_id: str
 
 class PostOutShort(BaseModel):
     id: str
     title: str
     subraddit: str
+
 
 class PostOutDetail(BaseModel):
     id: str
@@ -53,13 +55,16 @@ class EditPost(BaseModel):
     description: str
 
 
-
 @app.post('/api/posts', response_model=PostIn)
 def create_post(
     new_post: PostIn,
     content_queries: ContentQueries = Depends(),
+    # Authentication - are you signed in
+    account: dict = Depends(authenticator.get_current_account_data)
 ):
-    return content_queries.create_post(new_post)
+    if account['id'] is not None:
+        return content_queries.create_post(new_post)
+
 
 @app.get('/api/main/{id}', response_model=PostOutShort)
 def get_post_short_by_id(
@@ -79,15 +84,25 @@ def get_post_detail_by_id(
 def delete_post_by_id(
     id: str,
     content_queries: ContentQueries = Depends(),
+    account: dict = Depends(authenticator.get_current_account_data)
 ):
-    return content_queries.delete_post(id)
+    get = content_queries.get_post_by_id(id)
+    creator = get['user_id']
+    if account['id'] == creator:
+        return content_queries.delete_post(id)
+    else:
+        raise HTTPException(status_code=401, detail="not working")
 
 @app.put('/api/post/{id}')
 def update_post_by_id(
     id: str,
     description: EditPost,
     content_queries: ContentQueries = Depends(),
+    account: dict = Depends(authenticator.get_current_account_data)
 ):
-    return content_queries.edit_post(id, description)
-
-
+    get = content_queries.get_post_by_id(id)
+    creator = get['user_id']
+    if account['id'] == creator:
+        return content_queries.edit_post(id, description)
+    else:
+        raise HTTPException(status_code=401, detail="not working")
